refactor: add common attack module
