+++ src/modules/update_room/index.ts
@@ -1,6 +1,6 @@
 import { broadcastResponse } from "../../websocket_server";
 import { Rooms } from "../../store/rooms";
-import { createUpdateRoomResponse } from "../../utils/utils";
+import { createUpdateRoomResponse } from "../../utils/responses";
 
 export const updateRoom = (id: number) => {
   const roomData = Rooms.getRooms().filter((room) => room.players.length < 2).map((room) => ({
